Bug Incidence ortho a Zoom!=1

+++ src/uti_phgrm/CPP_C3DC.cpp
@@ -531,6 +531,7 @@ class cAppli_MPI2Mnt
          std::string			  mMasqImGlob;
          bool                     mDebug;
          bool                     mPurge;
+         bool        mUseTA;
          void ExeCom(const std::string & aCom);
 };
 
@@ -674,6 +675,7 @@ void cAppli_MPI2Mnt::DoMTD()
                           + mStrRep
                           + " DoMEC=0  Purge=true ZoomI=4 ZoomF=2  IncMax=1.0 " +
                           + " DirMEC=" + mDirMTD
+                          + " UseTA=" + ToString(mUseTA)
                           + " ZoomF=" + ToString(mDeZoom)
                        ;
 
@@ -694,7 +696,8 @@ cAppli_MPI2Mnt::cAppli_MPI2Mnt(int argc,char ** argv) :
     mDoOrtho     (false),
     mMasqImGlob (""),
     mDebug       (false),
-    mPurge       (true)
+    mPurge       (true),
+    mUseTA       (false)
 {
    ElInitArgMain
    (
@@ -708,6 +711,7 @@ cAppli_MPI2Mnt::cAppli_MPI2Mnt(int argc,char ** argv) :
                     << EAM(mDoOrtho,"DoOrtho",true,"Generate ortho photo,  def=false")
                     << EAM(mMasqImGlob,"MasqImGlob",true,"Global Masq for ortho: if used, give full name of masq (e.g. MasqGlob.tif) ",eSAM_IsExistFileRP)
                     << EAM(mDebug,"Debug",true,"Debug !!!",eSAM_InternalUse)
+                    << EAM(mUseTA,"UseTA",true,"Use TA as filter when exist (Def=false)",eSAM_InternalUse)
    );
 
    if (mDoOrtho && (!EAMIsInit(&mDoMnt))) mDoMnt = mDoOrtho;
@@ -721,7 +725,7 @@ cAppli_MPI2Mnt::cAppli_MPI2Mnt(int argc,char ** argv) :
    mSetIm = mICNM->Get(mCFPI->mStrPat);
 
 // Probleme d'incoherence et pas purgee !!!
-   if (mPurge)
+   if ((mPurge) && (!mDebug))
    {
       ELISE_fp::PurgeDirRecursif(mDirApp+mDirMTD);
       ELISE_fp::PurgeDirRecursif(mDirApp+mDirBasc);
